# public/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Storage API",
        "description": "This is documentation for API",
        "contact": {
            "email": "apiteam@wordnik.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "depot.esmc.info",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/depot_item_operations/{id}": {
            "get": {
                "tags": [
                    "Depot item operation"
                ],
                "summary": "Return one depot item operations",
                "description": "This is method for find one unit by ID",
                "operationId": "findUnit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of depot item operation",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/DepotItemOperation"
                        }
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Depot item operation"
                ],
                "summary": "Update unit",
                "description": "This is method update depot item operation by ID",
                "operationId": "updateDepotItemOperation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of depot item operation",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "item_id",
                        "in": "query",
                        "description": "ID of item",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "depot_item_id",
                        "in": "query",
                        "description": "ID of depot item",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "depot_id",
                        "in": "query",
                        "description": "ID of depot",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "in_progress",
                            "completed",
                            "rejected"
                        ]
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "change",
                            "move",
                            "correct"
                        ]
                    },
                    {
                        "name": "opposite_operation_id",
                        "in": "query",
                        "description": "Opposite operation ID",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/DepotItemOperation"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
        "/units/": {
            "get": {
                "tags": [
                    "Unit"
                ],
                "summary": "Return all units with pagination",
                "description": "This is method for find all units",
                "operationId": "findAllUnits",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/Unit"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Unit"
                ],
                "summary": "Create new unit",
                "description": "This is method create new unit",
                "operationId": "storeUnit",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of unit",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "designation",
                        "in": "query",
                        "description": "Designation of unit",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "decimal_symbol_count",
                        "in": "query",
                        "description": "Count symbols",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "min_value",
                        "in": "query",
                        "description": "-",
                        "required": false,
                        "type": "number"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/Unit"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/units/{id}": {
            "get": {
                "tags": [
                    "Unit"
                ],
                "summary": "Return one unit",
                "description": "This is method for find one unit by ID",
                "operationId": "findUnit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of item",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/Item"
                        }
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Unit"
                ],
                "summary": "Update unit",
                "description": "This is method update unit by ID",
                "operationId": "updateUnit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of unit",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of unit",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "designation",
                        "in": "query",
                        "description": "Designation of unit",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "decimal_symbol_count",
                        "in": "query",
                        "description": "Count symbols",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "min_value",
                        "in": "query",
                        "description": "-",
                        "required": false,
                        "type": "number"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/Unit"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Unit"
                ],
                "summary": "Remove unit by ID",
                "description": "This is method remove unit",
                "operationId": "destroyUnit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID unit for remove",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "bad request"
                    }
                }
            }
        }
    },
    "definitions": {
        "DepotItemOperation": {
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "item_id": {
                                "type": "integer"
                            },
                            "depot_id": {
                                "type": "integer"
                            },
                            "depot_item_id": {
                                "type": "integer"
                            },
                            "status": {
                                "type": "string"
                            },
                            "type": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "type": "object"
        },
...
        "Item": {
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "alias": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "description": {
                                "type": "string"
                            },
                            "group_id": {
                                "type": "integer"
                            },
                            "is_available": {
                                "type": "integer"
                            },
                            "created_at": {
                                "type": "dateTime"
                            },
                            "updated_at": {
                                "type": "dateTime"
                            },
                            "unit_id": {
                                "type": "integer"
                            }
                        }
                    }
                }
            },
            "type": "object"
        },
        "Unit": {
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "description": {
                                "type": "string"
                            },
                            "decimal_symbols_count": {
                                "type": "numeric"
                            }
                        }
                    }
                }
            },
            "type": "object"
        }
    },
...
}